Задание 8.1.4 - Изменил тип переменных на long

+++ Module_08/Quest_814/ComputeDisk.cs
@@ -5,9 +5,9 @@ namespace Module_08.Quest_814;
 public class ComputeDisk
 {
     public string Name { get; }
-    public double Space { get; }
-    public double EmptySpace { get; }
-    public ComputeDisk(string name, double space, double emptySpace){
+    public long Space { get; }
+    public long EmptySpace { get; }
+    public ComputeDisk(string name, long space, long emptySpace){
         Name = name;
         Space = space;
         EmptySpace = emptySpace;
